+++ apps/web/src/app/groups/[groupId]/expenses/expense-form.tsx
@@ -1,1949 +1,2 @@
-import { CategorySelector } from '@/components/category-selector'
-import { CurrencySelector } from '@/components/currency-selector'
-import { ExpenseDocumentsInput } from '@/components/expense-documents-input'
-import Link from '@/components/link'
-import { ParticipantDistributionFooter } from '@/components/participant-distribution-footer'
-import { ParticipantRowAmountPreview } from '@/components/participant-row-amount-preview'
-import { SubmitButton } from '@/components/submit-button'
-import { Button } from '@/components/ui/button'
-import {
-  Card,
-  CardContent,
-  CardDescription,
-  CardHeader,
-  CardTitle,
-} from '@/components/ui/card'
-import { Checkbox } from '@/components/ui/checkbox'
-import {
-  Collapsible,
-  CollapsibleContent,
-  CollapsibleTrigger,
-} from '@/components/ui/collapsible'
-import {
-  Form,
-  FormControl,
-  FormDescription,
-  FormField,
-  FormItem,
-  FormLabel,
-  FormMessage,
-} from '@/components/ui/form'
-import { Input } from '@/components/ui/input'
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from '@/components/ui/select'
-import { useLocale } from '@/i18n/react'
-import { Locale } from '@/i18n/request'
-import { randomId } from '@/lib/api'
-import { getCurrency, useCurrencies } from '@/lib/currency'
-import { RuntimeFeatureFlags } from '@/lib/featureFlags'
-import { useCurrencyRate } from '@/lib/hooks'
-import { expenseFormSchema, ExpenseFormValues } from '@/lib/schemas'
-import { calculatePaidByShare, calculateShare } from '@/lib/totals'
-import {
-  amountAsDecimal,
-  amountAsMinorUnits,
-  cn,
-  getCurrencyFromGroup,
-} from '@/lib/utils'
-import type { CreateExpenseSearch } from '@/router/schemas'
-import { trpc } from '@/trpc/client'
-import { zodResolver } from '@hookform/resolvers/zod'
-import type { AppRouterOutput } from '@spliit/api/router'
-import {
-  categoryIdSchema,
-  DEFAULT_CATEGORIES,
-  DEFAULT_CATEGORY_ID,
-  PAYMENT_CATEGORY_ID,
-  RecurrenceRule,
-} from '@spliit/domain'
-import { ArrowLeft, ChevronRight, Save } from 'lucide-react'
-import { useEffect, useState } from 'react'
-import type { Resolver } from 'react-hook-form'
-import { useForm } from 'react-hook-form'
-import { useTranslation } from 'react-i18next'
-import { match } from 'ts-pattern'
-import { DeletePopup } from '../../../../components/delete-popup'
-import { Textarea } from '../../../../components/ui/textarea'
-
-const enforceCurrencyPattern = (value: string) =>
-  value
-    .replace(/^\s*-/, '_') // replace leading minus with _
-    .replace(/[.,]/, '#') // replace first comma with #
-    .replace(/[-.,]/g, '') // remove other minus and commas characters
-    .replace(/_/, '-') // change back _ to minus
-    .replace(/#/, '.') // change back # to dot
-    .replace(/[^-\d.]/g, '') // remove all non-numeric characters
-
-const parseCategoryIdFromUrl = (raw: string | null | undefined) => {
-  if (!raw) return DEFAULT_CATEGORY_ID
-  const parsed = categoryIdSchema.safeParse(raw)
-  return parsed.success ? parsed.data : DEFAULT_CATEGORY_ID
-}
-
-const getDefaultSplittingOptions = (
-  group: NonNullable<AppRouterOutput['groups']['get']['group']>,
-) => {
-  // Default splitting: all ledger participants (active members + pending
-  // invitations), evenly. We no longer read per-device splitting defaults
-  // from localStorage; server-backed account preferences can replace this in
-  // a future pass.
-  return {
-    splitMode: 'EVENLY' as const,
-    paidFor: group.participants.map(({ id }) => ({
-      participant: id,
-      shares: '1' as any, // Use string to ensure consistent schema handling
-    })),
-  }
-}
-
-async function persistDefaultSplittingOptions(
-  _groupId: string,
-  _expenseFormValues: ExpenseFormValues,
-) {
-  // No-op: per-device splitting defaults were stored in localStorage before
-  // the account-backed product. Server-backed account preferences can replace
-  // this in a future pass.
-}
-
-export function ExpenseForm({
-  group,
-  expense,
-  searchParams,
-  onSubmit,
-  onDelete,
-  runtimeFeatureFlags,
-  currentLedgerParticipantId,
-  readOnly = false,
-}: {
-  group: NonNullable<AppRouterOutput['groups']['get']['group']>
-  expense?: AppRouterOutput['groups']['expenses']['get']['expense']
-  // Search-param defaults from the create route (reimbursement flow, receipt
-  // scan, etc.). Only provided when rendering the create form; undefined on
-  // the edit form, where defaults come from the loaded expense.
-  searchParams?: CreateExpenseSearch
-  onSubmit: (value: ExpenseFormValues) => Promise<void>
-  onDelete?: () => Promise<void>
-  runtimeFeatureFlags: RuntimeFeatureFlags
-  // Server-backed ledger participant id for the signed-in account. When
-  // creating an expense, we default the payer to this participant.
-  currentLedgerParticipantId?: string | null
-  readOnly?: boolean
-}) {
-  const { t } = useTranslation(undefined, { keyPrefix: 'ExpenseForm' })
-  const { t: tGroups } = useTranslation(undefined, { keyPrefix: 'Groups' })
-  const locale = useLocale() as Locale
-  const extractCategoryMutation = trpc.ai.extractCategoryFromTitle.useMutation()
-  const isCreate = expense === undefined
-  // Fall back to an empty object so the create-mode defaultValues below can
-  // still read optional fields when the create route happens to be rendered
-  // without any search params (e.g. a bare `/expenses/create` link).
-  const createSearch = searchParams ?? {}
-
-  const getSelectedRecurrenceRule = (field?: { value: string }) => {
-    return field?.value as RecurrenceRule
-  }
-  const defaultSplittingOptions = getDefaultSplittingOptions(group)
-  const groupCurrency = getCurrencyFromGroup(group)
-
-  // Default paidByList for the create path: a single row assigning the
-  // full amount to the signed-in user's ledger participant. The amount
-  // is in major units (same convention as the `amount` field).
-  const defaultPaidByList = currentLedgerParticipantId
-    ? [
-        {
-          participant: currentLedgerParticipantId,
-          shares: (Number(createSearch.amount) || 0) as any,
-        },
-      ]
-    : []
-
-  const form = useForm<ExpenseFormValues, any, ExpenseFormValues>({
-    resolver: zodResolver(expenseFormSchema) as Resolver<ExpenseFormValues>,
-    defaultValues: expense
-      ? {
-          title: expense.title,
-          expenseDate: expense.expenseDate ?? new Date(),
-          amount: amountAsDecimal(expense.amount, groupCurrency),
-          originalCurrency: expense.originalCurrency ?? group.currencyCode,
-          originalAmount:
-            expense.originalAmount != null
-              ? amountAsDecimal(
-                  expense.originalAmount,
-                  expense.originalCurrency
-                    ? (getCurrency(expense.originalCurrency) ?? groupCurrency)
-                    : groupCurrency,
-                )
-              : undefined,
-          conversionRate: expense.conversionRate?.toNumber(),
-          category: expense.categoryId,
-          paidBySplitMode: expense.paidBySplitMode,
-          paidByList: expense.paidByList.map(
-            ({ ledgerParticipantId, shares }) => {
-              // shares are stored in originalCurrency minor units when
-              // expense.originalCurrency is set; convert them to the matching
-              // currency's decimal value for the form input.
-              const rowCurrency = expense.originalCurrency
-                ? (getCurrency(expense.originalCurrency) ?? groupCurrency)
-                : groupCurrency
-              return {
-                participant: ledgerParticipantId,
-                shares: (expense.paidBySplitMode === 'BY_AMOUNT'
-                  ? amountAsDecimal(shares, rowCurrency)
-                  : (shares / 100).toString()) as any,
-              }
-            },
-          ),
-          isMultiPayer: expense.paidByList.length > 1,
-          paidFor: expense.paidFor.map(({ ledgerParticipantId, shares }) => ({
-            participant: ledgerParticipantId,
-            shares: (expense.splitMode === 'BY_AMOUNT'
-              ? amountAsDecimal(shares, groupCurrency)
-              : (shares / 100).toString()) as any,
-          })),
-          splitMode: expense.splitMode,
-          saveDefaultSplittingOptions: false,
-          isReimbursement: expense.isReimbursement,
-          documents: expense.documents,
-          notes: expense.notes ?? '',
-          recurrenceRule: expense.recurrenceRule ?? undefined,
-        }
-      : createSearch.reimbursement
-        ? {
-            title: t('reimbursement'),
-            expenseDate: new Date(),
-            amount: amountAsDecimal(
-              Number(createSearch.amount) || 0,
-              groupCurrency,
-            ),
-            originalCurrency: group.currencyCode,
-            originalAmount: undefined,
-            conversionRate: undefined,
-            category: PAYMENT_CATEGORY_ID,
-            paidBySplitMode: 'BY_AMOUNT' as const,
-            paidByList: createSearch.from
-              ? [
-                  {
-                    participant: createSearch.from,
-                    shares: amountAsDecimal(
-                      Number(createSearch.amount) || 0,
-                      groupCurrency,
-                    ) as any,
-                  },
-                ]
-              : [],
-            isMultiPayer: false,
-            paidFor: [
-              createSearch.to
-                ? {
-                    participant: createSearch.to,
-                    shares: '1' as any,
-                  }
-                : undefined,
-            ],
-            isReimbursement: true,
-            splitMode: defaultSplittingOptions.splitMode,
-            saveDefaultSplittingOptions: false,
-            documents: [],
-            notes: '',
-            recurrenceRule: RecurrenceRule.NONE,
-          }
-        : {
-            title: createSearch.title ?? '',
-            expenseDate: createSearch.date
-              ? new Date(createSearch.date)
-              : new Date(),
-            amount: Number(createSearch.amount) || 0,
-            originalCurrency: group.currencyCode ?? undefined,
-            originalAmount: undefined,
-            conversionRate: undefined,
-            category: parseCategoryIdFromUrl(createSearch.categoryId),
-            paidBySplitMode: 'BY_AMOUNT' as const,
-            paidByList: defaultPaidByList as any,
-            isMultiPayer: false,
-            paidFor: defaultSplittingOptions.paidFor,
-            isReimbursement: false,
-            splitMode: defaultSplittingOptions.splitMode,
-            saveDefaultSplittingOptions: false,
-            documents: createSearch.imageUrl
-              ? [
-                  {
-                    id: randomId(),
-                    url: createSearch.imageUrl,
-                    width: Number(createSearch.imageWidth),
-                    height: Number(createSearch.imageHeight),
-                  },
-                ]
-              : [],
-            notes: '',
-            recurrenceRule: RecurrenceRule.NONE,
-          },
-  })
-  const [isCategoryLoading, setCategoryLoading] = useState(false)
-
-  const submit = async (values: ExpenseFormValues) => {
-    if (readOnly) return
-    await persistDefaultSplittingOptions(group.id, values)
-
-    // Store monetary amounts in minor units (cents)
-    values.amount = amountAsMinorUnits(values.amount, groupCurrency)
-    values.paidFor = values.paidFor.map(({ participant, shares }) => ({
-      participant,
-      shares:
-        values.splitMode === 'BY_AMOUNT'
-          ? amountAsMinorUnits(shares, groupCurrency)
-          : shares,
-    }))
-    values.paidByList = values.paidByList.map(({ participant, shares }) => {
-      // paidByList.shares are entered in originalCurrency when set, so
-      // the minor-units conversion must use that currency, not the group
-      // currency — otherwise a EUR-group expense paid in USD would be
-      // stored as cents-of-EUR-denominated-dollars.
-      const payerCurrency =
-        values.originalCurrency && getCurrency(values.originalCurrency)
-          ? getCurrency(values.originalCurrency)!
-          : groupCurrency
-      return {
-        participant,
-        shares:
-          values.paidBySplitMode === 'BY_AMOUNT'
-            ? amountAsMinorUnits(shares, payerCurrency)
-            : shares,
-      }
-    })
-
-    // Convert originalAmount back to minor units for storage
-    if (values.originalAmount != null && values.originalCurrency) {
-      const origCurrency = getCurrency(values.originalCurrency) ?? groupCurrency
-      values.originalAmount = amountAsMinorUnits(
-        values.originalAmount,
-        origCurrency,
-      )
-    }
-
-    // Currency should be blank if same as group currency
-    if (!conversionRequired) {
-      delete values.originalAmount
-      delete values.originalCurrency
-    }
-    return onSubmit(values)
-  }
-
-  const [isIncome, setIsIncome] = useState(Number(form.getValues().amount) < 0)
-  const [manuallyEditedParticipants, setManuallyEditedParticipants] = useState<
-    Set<string>
-  >(new Set())
-  const [manuallyEditedPayers, setManuallyEditedPayers] = useState<Set<string>>(
-    new Set(),
-  )
-
-  const sExpense = isIncome ? 'Income' : 'Expense'
-
-  const originalCurrencyCode = form.getValues('originalCurrency')
-  const originalCurrency = originalCurrencyCode
-    ? (getCurrency(originalCurrencyCode) ?? {
-        code: '',
-        symbol: 'Custom',
-        rounding: 0,
-        decimal_digits: 2,
-      })
-    : { code: '', symbol: 'Custom', rounding: 0, decimal_digits: 2 }
-  const originalCurrencies = useCurrencies('')
-  const exchangeRate = useCurrencyRate(
-    form.watch('expenseDate'),
-    form.watch('originalCurrency') ?? '',
-    groupCurrency.code,
-  )
-
-  const conversionRequired =
-    group.currencyCode &&
-    group.currencyCode.length &&
-    originalCurrency.code.length &&
-    originalCurrency.code !== group.currencyCode
-
-  useEffect(() => {
-    setManuallyEditedParticipants(new Set())
-  }, [form.watch('splitMode'), form.watch('amount')])
-
-  useEffect(() => {
-    setManuallyEditedPayers(new Set())
-  }, [form.watch('paidBySplitMode'), form.watch('amount')])
-
-  useEffect(() => {
-    const splitMode = form.getValues().splitMode
-
-    // Only auto-balance for split mode 'Unevenly - By amount'
-    if (
-      splitMode === 'BY_AMOUNT' &&
-      (form.getFieldState('paidFor').isDirty ||
-        form.getFieldState('amount').isDirty)
-    ) {
-      const totalAmount = Number(form.getValues().amount) || 0
-      const paidFor = form.getValues().paidFor
-      let newPaidFor = [...paidFor]
-
-      const editedParticipants = Array.from(manuallyEditedParticipants)
-      let remainingAmount = totalAmount
-      let remainingParticipants = newPaidFor.length - editedParticipants.length
-
-      newPaidFor = newPaidFor.map((participant) => {
-        if (editedParticipants.includes(participant.participant)) {
-          const participantShare = Number(participant.shares) || 0
-          if (splitMode === 'BY_AMOUNT') {
-            remainingAmount -= participantShare
-          }
-          return participant
-        }
-        return participant
-      })
-
-      if (remainingParticipants > 0) {
-        let amountPerRemaining = 0
-        if (splitMode === 'BY_AMOUNT') {
-          amountPerRemaining = remainingAmount / remainingParticipants
-        }
-
-        newPaidFor = newPaidFor.map((participant) => {
-          if (!editedParticipants.includes(participant.participant)) {
-            return {
-              ...participant,
-              shares: amountPerRemaining.toFixed(
-                groupCurrency.decimal_digits,
-              ) as any, // Keep as string for consistent schema handling
-            }
-          }
-          return participant
-        })
-      }
-      form.setValue('paidFor', newPaidFor, { shouldValidate: true })
-    }
-  }, [
-    manuallyEditedParticipants,
-    form.watch('amount'),
-    form.watch('splitMode'),
-  ])
-
-  useEffect(() => {
-    const splitMode = form.getValues().paidBySplitMode
-
-    if (
-      splitMode === 'BY_AMOUNT' &&
-      (form.getFieldState('paidByList').isDirty ||
-        form.getFieldState('amount').isDirty)
-    ) {
-      // Distribute against originalAmount (in originalCurrency) when set,
-      // else against amount (group currency), so shares stay consistent
-      // with the schema sum-check and the Total footer.
-      const originalCurrencyCode = form.getValues('originalCurrency')
-      const payerCurrency = originalCurrencyCode
-        ? (getCurrency(originalCurrencyCode) ?? groupCurrency)
-        : groupCurrency
-      const totalAmount =
-        originalCurrencyCode && form.getValues('originalAmount') != null
-          ? Number(form.getValues('originalAmount')) || 0
-          : Number(form.getValues().amount) || 0
-
-      const paidByList = form.getValues().paidByList
-      let newPaidByList = [...paidByList]
-
-      const editedPayers = Array.from(manuallyEditedPayers)
-      let remainingAmount = totalAmount
-      let remainingPayers = newPaidByList.length - editedPayers.length
-
-      newPaidByList = newPaidByList.map((payer) => {
-        if (editedPayers.includes(payer.participant)) {
-          const payerShare = Number(payer.shares) || 0
-          remainingAmount -= payerShare
-          return payer
-        }
-        return payer
-      })
-
-      if (remainingPayers > 0) {
-        const amountPerRemaining = remainingAmount / remainingPayers
-        newPaidByList = newPaidByList.map((payer) => {
-          if (!editedPayers.includes(payer.participant)) {
-            return {
-              ...payer,
-              shares: amountPerRemaining.toFixed(
-                payerCurrency.decimal_digits,
-              ) as any,
-            }
-          }
-          return payer
-        })
-      }
-      form.setValue('paidByList', newPaidByList, { shouldValidate: true })
-    }
-  }, [
-    manuallyEditedPayers,
-    form.watch('amount'),
-    form.watch('paidBySplitMode'),
-    form.watch('originalCurrency'),
-    form.watch('originalAmount'),
-  ])
-
-  const [usingCustomConversionRate, setUsingCustomConversionRate] = useState(
-    !!form.formState.defaultValues?.conversionRate,
-  )
-
-  useEffect(() => {
-    if (!usingCustomConversionRate && exchangeRate.data) {
-      form.setValue('conversionRate', exchangeRate.data)
-    }
-  }, [exchangeRate.data, usingCustomConversionRate])
-
-  useEffect(() => {
-    if (!form.getFieldState('originalAmount').isTouched) return
-    const originalAmount = form.getValues('originalAmount') ?? 0
-    const conversionRate = form.getValues('conversionRate')
-
-    if (conversionRate && originalAmount) {
-      const rate = Number(conversionRate)
-      const convertedAmount = originalAmount * rate
-      if (!Number.isNaN(convertedAmount)) {
-        const v = enforceCurrencyPattern(
-          convertedAmount.toFixed(groupCurrency.decimal_digits),
-        )
-        const income = Number(v) < 0
-        setIsIncome(income)
-        if (income) form.setValue('isReimbursement', false)
-        form.setValue('amount', Number(v))
-      }
-    }
-  }, [
-    form.watch('originalAmount'),
-    form.watch('conversionRate'),
-    form.getFieldState('originalAmount').isTouched,
-  ])
-
-  let conversionRateMessage = ''
-  if (exchangeRate.isLoading) {
-    conversionRateMessage = t('conversionRateState.loading')
-  } else {
-    let ratesDisplay = ''
-    if (exchangeRate.data) {
-      // non breaking spaces so the rate text is not split with line feeds
-      ratesDisplay = `${form.getValues('originalCurrency')}\xa01\xa0=\xa0${
-        group.currencyCode
-      }\xa0${exchangeRate.data}`
-    }
-    if (exchangeRate.error) {
-      if (exchangeRate.error instanceof RangeError && exchangeRate.data)
-        conversionRateMessage = t('conversionRateState.dateMismatch', {
-          date: exchangeRate.error.message,
-        })
-      else {
-        conversionRateMessage = t('conversionRateState.error')
-      }
-      conversionRateMessage +=
-        ' ' +
-        (ratesDisplay.length
-          ? `${t('conversionRateState.staleRate')} ${ratesDisplay}`
-          : t('conversionRateState.noRate'))
-    } else {
-      conversionRateMessage = ratesDisplay.length
-        ? `${t('conversionRateState.success')} ${ratesDisplay}`
-        : t('conversionRateState.currencyNotFound')
-    }
-  }
-
-  return (
-    <Form {...form}>
-      <form onSubmit={form.handleSubmit(submit)}>
-        {readOnly && (
-          <p className="text-sm text-muted-foreground mb-4">
-            {t('readOnlyNotice')}
-          </p>
-        )}
-        <Card>
-          <CardHeader className="flex flex-row items-center gap-2 space-y-0">
-            <Button variant="ghost" size="icon" asChild className="-ml-2">
-              <Link
-                href={`/groups/${group.id}/expenses`}
-                title={tGroups('backToExpenses')}
-              >
-                <ArrowLeft className="w-4 h-4" />
-              </Link>
-            </Button>
-            <CardTitle>
-              {t(`${sExpense}.${isCreate ? 'create' : 'edit'}`)}
-            </CardTitle>
-          </CardHeader>
-          <CardContent className="grid sm:grid-cols-2 gap-6">
-            <FormField
-              control={form.control}
-              name="title"
-              render={({ field }) => (
-                <FormItem className="">
-                  <FormLabel>{t(`${sExpense}.TitleField.label`)}</FormLabel>
-                  <FormControl>
-                    <Input
-                      placeholder={t(`${sExpense}.TitleField.placeholder`)}
-                      className="text-base"
-                      disabled={readOnly}
-                      {...field}
-                      onBlur={async () => {
-                        field.onBlur() // avoid skipping other blur event listeners since we overwrite `field`
-                        if (
-                          !readOnly &&
-                          runtimeFeatureFlags.enableCategoryExtract
-                        ) {
-                          setCategoryLoading(true)
-                          const { categoryId } =
-                            await extractCategoryMutation.mutateAsync({
-                              description: field.value,
-                            })
-                          form.setValue('category', categoryId)
-                          setCategoryLoading(false)
-                        }
-                      }}
-                    />
-                  </FormControl>
-                  <FormDescription>
-                    {t(`${sExpense}.TitleField.description`)}
-                  </FormDescription>
-                  <FormMessage />
-                </FormItem>
-              )}
-            />
-
-            <FormField
-              control={form.control}
-              name="expenseDate"
-              render={({ field }) => (
-                <FormItem className="sm:order-1">
-                  <FormLabel>{t(`${sExpense}.DateField.label`)}</FormLabel>
-                  <FormControl>
-                    <Input
-                      className="date-base"
-                      type="date"
-                      defaultValue={formatDate(field.value)}
-                      disabled={readOnly}
-                      onChange={(event) => {
-                        return field.onChange(new Date(event.target.value))
-                      }}
-                    />
-                  </FormControl>
-                  <FormDescription>
-                    {t(`${sExpense}.DateField.description`)}
-                  </FormDescription>
-                  <FormMessage />
-                </FormItem>
-              )}
-            />
-
-            <FormField
-              name="originalCurrency"
-              render={({ field: { onChange, ...field } }) => (
-                <FormItem className="sm:order-3">
-                  <FormLabel>{t(`${sExpense}.currencyField.label`)}</FormLabel>
-                  <FormControl>
-                    {group.currencyCode ? (
-                      <CurrencySelector
-                        currencies={originalCurrencies}
-                        defaultValue={form.watch(field.name) ?? ''}
-                        isLoading={false}
-                        disabled={readOnly}
-                        onValueChange={(v) => onChange(v)}
-                      />
-                    ) : (
-                      <Input
-                        className="text-base"
-                        disabled={true}
-                        {...field}
-                        placeholder={group.currency}
-                      />
-                    )}
-                  </FormControl>
-                  <FormDescription>
-                    {t(`${sExpense}.currencyField.description`)}{' '}
-                    {!group.currencyCode && t('conversionUnavailable')}
-                  </FormDescription>
-                  <FormMessage />
-                </FormItem>
-              )}
-            />
-
-            <div
-              className={`sm:order-4 ${
-                !conversionRequired ? 'max-sm:hidden sm:invisible' : ''
-              } col-span-2 md:col-span-1 space-y-2`}
-            >
-              <FormField
-                control={form.control}
-                name="originalAmount"
-                render={({ field: { onChange, ...field } }) => (
-                  <FormItem>
-                    <FormLabel>{t('originalAmountField.label')}</FormLabel>
-                    <div className="flex items-baseline gap-2">
-                      <span>{originalCurrency.symbol}</span>
-                      <FormControl>
-                        <Input
-                          className="text-base max-w-[120px]"
-                          type="text"
-                          inputMode="decimal"
-                          placeholder="0.00"
-                          disabled={readOnly}
-                          onChange={(event) => {
-                            const v = enforceCurrencyPattern(event.target.value)
-                            onChange(v)
-                          }}
-                          {...field}
-                          onFocus={(e) => {
-                            const target = e.currentTarget
-                            setTimeout(() => target.select(), 1)
-                          }}
-                        />
-                      </FormControl>
-                    </div>
-                    <FormDescription>
-                      {isNaN(form.getValues('expenseDate').getTime()) ? (
-                        t('conversionRateState.noDate')
-                      ) : form.getValues('expenseDate') &&
-                        !usingCustomConversionRate ? (
-                        <>
-                          {conversionRateMessage}
-                          {!exchangeRate.isLoading && (
-                            <Button
-                              className="h-auto py-0"
-                              variant="link"
-                              onClick={() => exchangeRate.refresh()}
-                              disabled={readOnly}
-                            >
-                              {t('conversionRateState.refresh')}
-                            </Button>
-                          )}
-                        </>
-                      ) : (
-                        t('conversionRateState.customRate')
-                      )}
-                    </FormDescription>
-                    <FormMessage />
-                  </FormItem>
-                )}
-              />
-              <Collapsible
-                open={usingCustomConversionRate}
-                onOpenChange={setUsingCustomConversionRate}
-              >
-                <CollapsibleTrigger asChild>
-                  <Button variant="link" className="-mx-4" disabled={readOnly}>
-                    {usingCustomConversionRate
-                      ? t('conversionRateField.useApi')
-                      : t('conversionRateField.useCustom')}
-                  </Button>
-                </CollapsibleTrigger>
-                <CollapsibleContent>
-                  <FormField
-                    control={form.control}
-                    name="conversionRate"
-                    render={({ field: { onChange, ...field } }) => (
-                      <FormItem
-                        className={`sm:order-4 ${
-                          !conversionRequired
-                            ? 'max-sm:hidden sm:invisible'
-                            : ''
-                        }`}
-                      >
-                        <FormLabel>{t('conversionRateField.label')}</FormLabel>
-                        <div className="flex items-baseline gap-2">
-                          <span>
-                            {originalCurrency.symbol} 1 = {group.currency}
-                          </span>
-                          <FormControl>
-                            <Input
-                              className="text-base max-w-[120px]"
-                              type="text"
-                              inputMode="decimal"
-                              placeholder="0.00"
-                              disabled={readOnly}
-                              onChange={(event) => {
-                                const v = enforceCurrencyPattern(
-                                  event.target.value,
-                                )
-                                onChange(v)
-                              }}
-                              {...field}
-                              onFocus={(e) => {
-                                const target = e.currentTarget
-                                setTimeout(() => target.select(), 1)
-                              }}
-                            />
-                          </FormControl>
-                        </div>
-                        <FormMessage />
-                      </FormItem>
-                    )}
-                  />
-                </CollapsibleContent>
-              </Collapsible>
-            </div>
-            <FormField
-              control={form.control}
-              name="category"
-              render={({ field }) => (
-                <FormItem className="order-3 sm:order-2">
-                  <FormLabel>{t('categoryField.label')}</FormLabel>
-                  <CategorySelector
-                    categories={DEFAULT_CATEGORIES}
-                    defaultValue={
-                      form.watch(field.name) // may be overwritten externally
-                    }
-                    onValueChange={field.onChange}
-                    isLoading={isCategoryLoading}
-                    disabled={readOnly}
-                  />
-                  <FormDescription>
-                    {t(`${sExpense}.categoryFieldDescription`)}
-                  </FormDescription>
-                  <FormMessage />
-                </FormItem>
-              )}
-            />
-
-            <FormField
-              control={form.control}
-              name="amount"
-              render={({ field: { onChange, ...field } }) => (
-                <FormItem className="sm:order-5">
-                  <FormLabel>{t('amountField.label')}</FormLabel>
-                  <div className="flex items-baseline gap-2">
-                    <span>{group.currency}</span>
-                    <FormControl>
-                      <Input
-                        className="text-base max-w-[120px]"
-                        type="text"
-                        inputMode="decimal"
-                        placeholder="0.00"
-                        disabled={readOnly}
-                        onChange={(event) => {
-                          const v = enforceCurrencyPattern(event.target.value)
-                          const income = Number(v) < 0
-                          setIsIncome(income)
-                          if (income) form.setValue('isReimbursement', false)
-                          onChange(v)
-                        }}
-                        onFocus={(e) => {
-                          // we're adding a small delay to get around safaris issue with onMouseUp deselecting things again
-                          const target = e.currentTarget
-                          setTimeout(() => target.select(), 1)
-                        }}
-                        {...field}
-                      />
-                    </FormControl>
-                  </div>
-                  <FormMessage />
-
-                  {!isIncome && (
-                    <FormField
-                      control={form.control}
-                      name="isReimbursement"
-                      render={({ field }) => (
-                        <FormItem className="flex flex-row gap-2 items-center space-y-0 pt-2">
-                          <FormControl>
-                            <Checkbox
-                              checked={field.value}
-                              onCheckedChange={field.onChange}
-                              disabled={readOnly}
-                            />
-                          </FormControl>
-                          <div>
-                            <FormLabel>
-                              {t('isReimbursementField.label')}
-                            </FormLabel>
-                          </div>
-                        </FormItem>
-                      )}
-                    />
-                  )}
-                </FormItem>
-              )}
-            />
-
-            <FormField
-              control={form.control}
-              name="notes"
-              render={({ field }) => (
-                <FormItem className="sm:order-6">
-                  <FormLabel>{t('notesField.label')}</FormLabel>
-                  <FormControl>
-                    <Textarea
-                      className="text-base"
-                      disabled={readOnly}
-                      {...field}
-                    />
-                  </FormControl>
-                </FormItem>
-              )}
-            />
-            <FormField
-              control={form.control}
-              name="recurrenceRule"
-              render={({ field }) => (
-                <FormItem className="sm:order-5">
-                  <FormLabel>{t('Expense.recurrenceRule.label')}</FormLabel>
-                  <Select
-                    onValueChange={(value) => {
-                      form.setValue('recurrenceRule', value as RecurrenceRule)
-                    }}
-                    defaultValue={getSelectedRecurrenceRule(field)}
-                    disabled={readOnly}
-                  >
-                    <SelectTrigger>
-                      <SelectValue placeholder="NONE" />
-                    </SelectTrigger>
-                    <SelectContent>
-                      <SelectItem value="NONE">
-                        {t('Expense.recurrenceRule.none')}
-                      </SelectItem>
-                      <SelectItem value="DAILY">
-                        {t('Expense.recurrenceRule.daily')}
-                      </SelectItem>
-                      <SelectItem value="WEEKLY">
-                        {t('Expense.recurrenceRule.weekly')}
-                      </SelectItem>
-                      <SelectItem value="MONTHLY">
-                        {t('Expense.recurrenceRule.monthly')}
-                      </SelectItem>
-                    </SelectContent>
-                  </Select>
-                  <FormDescription>
-                    {t('Expense.recurrenceRule.description')}
-                  </FormDescription>
-                  <FormMessage />
-                </FormItem>
-              )}
-            />
-          </CardContent>
-        </Card>
-
-        <Card className="mt-4">
-          <CardHeader>
-            <CardTitle className="flex justify-between">
-              <span>{t(`${sExpense}.paidByField.label`)}</span>
-              {form.watch('isMultiPayer') && (
-                <Button
-                  variant="link"
-                  type="button"
-                  className="-my-2 -mx-4"
-                  disabled={readOnly}
-                  onClick={() => {
-                    const paidByList = form.getValues().paidByList
-                    const allSelected =
-                      paidByList.length === group.participants.length
-                    const newPaidByList = allSelected
-                      ? []
-                      : group.participants.map((p) => ({
-                          participant: p.id,
-                          shares: (paidByList.find(
-                            (pb) => pb.participant === p.id,
-                          )?.shares ?? '1') as any,
-                        }))
-                    form.setValue('paidByList', newPaidByList as any, {
-                      shouldDirty: true,
-                      shouldTouch: true,
-                      shouldValidate: true,
-                    })
-                  }}
-                >
-                  {form.getValues().paidByList.length ===
-                  group.participants.length ? (
-                    <>{t('selectNone')}</>
-                  ) : (
-                    <>{t('selectAll')}</>
-                  )}
-                </Button>
-              )}
-            </CardTitle>
-            <CardDescription>
-              {t(`${sExpense}.paidByField.description`)}
-            </CardDescription>
-          </CardHeader>
-          <CardContent>
-            {form.watch('isMultiPayer') ? (
-              (() => {
-                // When a non-group originalCurrency is set, payer shares are
-                // entered in that currency (matching `originalAmount`); the
-                // Total footer and per-row label use `payerCurrency` instead
-                // of `groupCurrency` so the form stays in one currency at a
-                // time. Falls back to the group currency when no conversion
-                // is set.
-                const originalCurrencyCode = form.watch('originalCurrency')
-                const originalPayerCurrency = originalCurrencyCode
-                  ? getCurrency(originalCurrencyCode)
-                  : undefined
-                const payerCurrency = originalPayerCurrency ?? groupCurrency
-                const isOriginalPayer = !!originalPayerCurrency
-                const targetAmount = isOriginalPayer
-                  ? Number(form.watch('originalAmount')) || 0
-                  : Number(form.watch('amount')) || 0
-                // Build a synthetic TotalsExpense so calculatePaidByShare
-                // returns each row's resolved share in payerCurrency cents.
-                // Shares are in major units in the form; BY_AMOUNT is
-                // converted to minor units, BY_PERCENTAGE to basis points.
-                // No originalCurrency/conversionRate is passed so the result
-                // stays in the input unit (payerCurrency) regardless of
-                // whether the ledger has a different currency.
-                const pbSplitMode = form.watch('paidBySplitMode')
-                const pbList = form.watch('paidByList')
-                const paidByListForCalc = pbList.map((p) => {
-                  const rawShares = Number(p.shares) || 0
-                  const shares =
-                    pbSplitMode === 'BY_PERCENTAGE'
-                      ? rawShares * 100
-                      : pbSplitMode === 'BY_AMOUNT'
-                        ? amountAsMinorUnits(rawShares, payerCurrency)
-                        : rawShares
-                  return { participant: { id: p.participant }, shares }
-                })
-                const paidByExpenseForCalc = {
-                  amount: amountAsMinorUnits(targetAmount, payerCurrency),
-                  paidByList: paidByListForCalc,
-                  paidBySplitMode: pbSplitMode,
-                  isReimbursement: false,
-                }
-                return (
-                  <>
-                    <FormField
-                      control={form.control}
-                      name="paidByList"
-                      render={() => (
-                        <FormItem className="sm:order-4 row-span-2 space-y-0">
-                          {group.participants.map((participant) => (
-                            <FormField
-                              key={participant.id}
-                              control={form.control}
-                              name="paidByList"
-                              render={({ field }) => {
-                                const { id, name, pending } = participant
-                                return (
-                                  <div
-                                    data-id={`${id}/${form.getValues().paidBySplitMode}/${payerCurrency.code}`}
-                                    className="flex flex-wrap gap-y-4 items-center border-t last-of-type:border-b last-of-type:!mb-4 -mx-6 px-6 py-3"
-                                  >
-                                    <FormItem className="flex-1 flex flex-row items-start space-x-3 space-y-0">
-                                      <FormControl>
-                                        <Checkbox
-                                          checked={field.value?.some(
-                                            ({ participant }) =>
-                                              participant === id,
-                                          )}
-                                          onCheckedChange={(checked) => {
-                                            if (readOnly) return
-                                            const options = {
-                                              shouldDirty: true,
-                                              shouldTouch: true,
-                                              shouldValidate: true,
-                                            }
-                                            checked
-                                              ? form.setValue(
-                                                  'paidByList',
-                                                  [
-                                                    ...field.value,
-                                                    {
-                                                      participant: id,
-                                                      shares: '1',
-                                                    },
-                                                  ] as any,
-                                                  options,
-                                                )
-                                              : form.setValue(
-                                                  'paidByList',
-                                                  field.value?.filter(
-                                                    (value) =>
-                                                      value.participant !== id,
-                                                  ),
-                                                  options,
-                                                )
-                                          }}
-                                          disabled={readOnly}
-                                        />
-                                      </FormControl>
-                                      <FormLabel className="text-sm font-normal flex-1">
-                                        {name}
-                                        {pending && (
-                                          <span className="ml-2 text-xs text-muted-foreground">
-                                            {t('participant.pending')}
-                                          </span>
-                                        )}
-                                        {field.value?.some(
-                                          ({ participant }) =>
-                                            participant === id,
-                                        ) &&
-                                          !form.watch('isReimbursement') && (
-                                            <ParticipantRowAmountPreview
-                                              amount={calculatePaidByShare(
-                                                id,
-                                                paidByExpenseForCalc,
-                                              )}
-                                              currency={payerCurrency}
-                                            />
-                                          )}
-                                      </FormLabel>
-                                    </FormItem>
-                                    <div className="flex">
-                                      {form.getValues().paidBySplitMode !==
-                                        'EVENLY' && (
-                                        <FormField
-                                          name={`paidByList[${field.value.findIndex(({ participant }) => participant === id)}].shares`}
-                                          render={() => {
-                                            const sharesLabel = (
-                                              <span
-                                                className={cn('text-sm', {
-                                                  'text-muted':
-                                                    !field.value?.some(
-                                                      ({ participant }) =>
-                                                        participant === id,
-                                                    ),
-                                                })}
-                                              >
-                                                {match(
-                                                  form.getValues()
-                                                    .paidBySplitMode,
-                                                )
-                                                  .with('BY_SHARES', () => (
-                                                    <>{t('shares')}</>
-                                                  ))
-                                                  .with('BY_PERCENTAGE', () => (
-                                                    <>%</>
-                                                  ))
-                                                  .with('BY_AMOUNT', () => (
-                                                    <>{payerCurrency.symbol}</>
-                                                  ))
-                                                  .otherwise(() => (
-                                                    <></>
-                                                  ))}
-                                              </span>
-                                            )
-                                            return (
-                                              <div>
-                                                <div className="flex gap-1 items-center">
-                                                  {form.getValues()
-                                                    .paidBySplitMode ===
-                                                    'BY_AMOUNT' && sharesLabel}
-                                                  <FormControl>
-                                                    <Input
-                                                      key={String(
-                                                        !field.value?.some(
-                                                          ({ participant }) =>
-                                                            participant === id,
-                                                        ),
-                                                      )}
-                                                      className="text-base w-[80px] -my-2"
-                                                      type="text"
-                                                      disabled={
-                                                        readOnly ||
-                                                        !field.value?.some(
-                                                          ({ participant }) =>
-                                                            participant === id,
-                                                        )
-                                                      }
-                                                      value={
-                                                        field.value?.find(
-                                                          ({ participant }) =>
-                                                            participant === id,
-                                                        )?.shares
-                                                      }
-                                                      onChange={(event) => {
-                                                        field.onChange(
-                                                          field.value.map(
-                                                            (p) =>
-                                                              p.participant ===
-                                                              id
-                                                                ? {
-                                                                    participant:
-                                                                      id,
-                                                                    shares:
-                                                                      enforceCurrencyPattern(
-                                                                        event
-                                                                          .target
-                                                                          .value,
-                                                                      ),
-                                                                  }
-                                                                : p,
-                                                          ),
-                                                        )
-                                                        setManuallyEditedPayers(
-                                                          (prev) =>
-                                                            new Set(prev).add(
-                                                              id,
-                                                            ),
-                                                        )
-                                                      }}
-                                                      inputMode={
-                                                        form.getValues()
-                                                          .paidBySplitMode ===
-                                                        'BY_AMOUNT'
-                                                          ? 'decimal'
-                                                          : 'numeric'
-                                                      }
-                                                      step={
-                                                        form.getValues()
-                                                          .paidBySplitMode ===
-                                                        'BY_AMOUNT'
-                                                          ? 10 **
-                                                            -payerCurrency.decimal_digits
-                                                          : 1
-                                                      }
-                                                    />
-                                                  </FormControl>
-                                                  {[
-                                                    'BY_SHARES',
-                                                    'BY_PERCENTAGE',
-                                                  ].includes(
-                                                    form.getValues()
-                                                      .paidBySplitMode,
-                                                  ) && sharesLabel}
-                                                </div>
-                                                <FormMessage className="float-right" />
-                                              </div>
-                                            )
-                                          }}
-                                        />
-                                      )}
-                                    </div>
-                                  </div>
-                                )
-                              }}
-                            />
-                          ))}
-                          <FormMessage />
-                        </FormItem>
-                      )}
-                    />
-
-                    {(() => {
-                      const pbSplitMode = form.watch('paidBySplitMode')
-                      const pbList = form.watch('paidByList')
-                      const paidByCount = pbList.length
-                      const sharesForFooter =
-                        pbSplitMode === 'BY_AMOUNT'
-                          ? pbList.map((p) =>
-                              amountAsMinorUnits(
-                                Number(p.shares) || 0,
-                                payerCurrency,
-                              ),
-                            )
-                          : pbSplitMode === 'BY_PERCENTAGE'
-                            ? pbList.map((p) => Number(p.shares) || 0)
-                            : pbList.map((p) => Number(p.shares) || 0)
-                      const targetForFooter =
-                        pbSplitMode === 'BY_PERCENTAGE'
-                          ? 100
-                          : amountAsMinorUnits(
-                              Number(targetAmount) || 0,
-                              payerCurrency,
-                            )
-                      return (
-                        <ParticipantDistributionFooter
-                          splitMode={pbSplitMode}
-                          targetAmount={targetForFooter}
-                          shares={sharesForFooter}
-                          currency={payerCurrency}
-                          paidByCount={paidByCount}
-                          dataTestId="paid-by-distribution-footer"
-                        />
-                      )
-                    })()}
-
-                    <Collapsible
-                      className="mt-5"
-                      defaultOpen={
-                        form.getValues().paidBySplitMode !== 'EVENLY'
-                      }
-                    >
-                      <CollapsibleTrigger asChild>
-                        <Button
-                          variant="link"
-                          className="-mx-4"
-                          disabled={readOnly}
-                        >
-                          {t('paidByAdvancedOptions')}
-                        </Button>
-                      </CollapsibleTrigger>
-                      <CollapsibleContent>
-                        <div className="grid sm:grid-cols-2 gap-6 pt-3">
-                          <FormField
-                            control={form.control}
-                            name="paidBySplitMode"
-                            render={({ field }) => (
-                              <FormItem>
-                                <FormLabel>
-                                  {t('PaidBySplitModeField.label')}
-                                </FormLabel>
-                                <FormControl>
-                                  <Select
-                                    onValueChange={(value) => {
-                                      form.setValue(
-                                        'paidBySplitMode',
-                                        value as any,
-                                        {
-                                          shouldDirty: true,
-                                          shouldTouch: true,
-                                          shouldValidate: true,
-                                        },
-                                      )
-                                    }}
-                                    defaultValue={field.value}
-                                    disabled={readOnly}
-                                  >
-                                    <SelectTrigger>
-                                      <SelectValue />
-                                    </SelectTrigger>
-                                    <SelectContent>
-                                      <SelectItem value="EVENLY">
-                                        {t('SplitModeField.evenly')}
-                                      </SelectItem>
-                                      <SelectItem value="BY_SHARES">
-                                        {t('SplitModeField.byShares')}
-                                      </SelectItem>
-                                      <SelectItem value="BY_PERCENTAGE">
-                                        {t('SplitModeField.byPercentage')}
-                                      </SelectItem>
-                                      <SelectItem value="BY_AMOUNT">
-                                        {t('SplitModeField.byAmount')}
-                                      </SelectItem>
-                                    </SelectContent>
-                                  </Select>
-                                </FormControl>
-                                <FormDescription>
-                                  {t(`${sExpense}.splitModeDescription`)}
-                                </FormDescription>
-                              </FormItem>
-                            )}
-                          />
-                        </div>
-                      </CollapsibleContent>
-                    </Collapsible>
-                  </>
-                )
-              })()
-            ) : (
-              <FormField
-                control={form.control}
-                name="paidByList"
-                render={() => {
-                  const paidByList = form.watch('paidByList')
-                  const selectedPayer = paidByList[0]?.participant ?? ''
-                  const amount = String(Number(form.watch('amount')) || 0)
-                  return (
-                    <FormItem>
-                      <Select
-                        value={selectedPayer}
-                        onValueChange={(value) => {
-                          form.setValue(
-                            'paidByList',
-                            [{ participant: value, shares: amount }] as any,
-                            {
-                              shouldDirty: true,
-                              shouldTouch: true,
-                              shouldValidate: true,
-                            },
-                          )
-                        }}
-                        disabled={readOnly}
-                      >
-                        <SelectTrigger>
-                          <SelectValue
-                            placeholder={
-                              t('Expense.paidByField.placeholder') as string
-                            }
-                          />
-                        </SelectTrigger>
-                        <SelectContent>
-                          {group.participants.map(({ id, name, pending }) => (
-                            <SelectItem key={id} value={id}>
-                              {name}
-                              {pending && (
-                                <span className="ml-2 text-xs text-muted-foreground">
-                                  {t('participant.pending')}
-                                </span>
-                              )}
-                            </SelectItem>
-                          ))}
-                        </SelectContent>
-                      </Select>
-                      <FormMessage />
-                    </FormItem>
-                  )
-                }}
-              />
-            )}
-
-            <FormField
-              control={form.control}
-              name="isMultiPayer"
-              render={({ field }) => (
-                <FormItem className="flex flex-row gap-2 items-center space-y-0 pt-4">
-                  <FormControl>
-                    <Checkbox
-                      checked={field.value}
-                      onCheckedChange={(checked) => {
-                        // Mirror Save-as-default style: settle on a single
-                        // payer row (shares = amount) when toggling either
-                        // direction so the multi-payer view shows the
-                        // current single selection already checked, and
-                        // collapsing back to single-payer keeps the first
-                        // selected row.
-                        const currentList = form.getValues().paidByList
-                        const amount = String(Number(form.watch('amount')) || 0)
-                        const firstParticipant =
-                          currentList[0]?.participant ??
-                          group.participants[0]?.id
-                        if (firstParticipant) {
-                          form.setValue(
-                            'paidByList',
-                            [
-                              {
-                                participant: firstParticipant,
-                                shares: amount,
-                              },
-                            ] as any,
-                            {
-                              shouldDirty: true,
-                              shouldValidate: true,
-                            },
-                          )
-                        }
-                        field.onChange(checked)
-                      }}
-                      disabled={readOnly}
-                    />
-                  </FormControl>
-                  <div>
-                    <FormLabel>{t('paidByAdvancedOptions')}</FormLabel>
-                  </div>
-                </FormItem>
-              )}
-            />
-          </CardContent>
-        </Card>
-
-        <Card className="mt-4">
-          <CardHeader>
-            <CardTitle className="flex justify-between">
-              <span>{t(`${sExpense}.paidFor.title`)}</span>
-              <Button
-                variant="link"
-                type="button"
-                className="-my-2 -mx-4"
-                disabled={readOnly}
-                onClick={() => {
-                  const paidFor = form.getValues().paidFor
-                  const allSelected =
-                    paidFor.length === group.participants.length
-                  const newPaidFor = allSelected
-                    ? []
-                    : group.participants.map((p) => ({
-                        participant: p.id,
-                        shares: (paidFor.find(
-                          (pfor) => pfor.participant === p.id,
-                        )?.shares ?? '1') as any, // Use string to ensure consistent schema handling
-                      }))
-                  form.setValue('paidFor', newPaidFor as any, {
-                    shouldDirty: true,
-                    shouldTouch: true,
-                    shouldValidate: true,
-                  })
-                }}
-              >
-                {form.getValues().paidFor.length ===
-                group.participants.length ? (
-                  <>{t('selectNone')}</>
-                ) : (
-                  <>{t('selectAll')}</>
-                )}
-              </Button>
-            </CardTitle>
-            <CardDescription>
-              {t(`${sExpense}.paidFor.description`)}
-            </CardDescription>
-          </CardHeader>
-          <CardContent>
-            <FormField
-              control={form.control}
-              name="paidFor"
-              render={() => (
-                <FormItem className="sm:order-4 row-span-2 space-y-0">
-                  {group.participants.map((participant) => (
-                    <FormField
-                      key={participant.id}
-                      control={form.control}
-                      name="paidFor"
-                      render={({ field }) => {
-                        const { id, name, pending } = participant
-                        return (
-                          <div
-                            data-id={`${id}/${form.getValues().splitMode}/${
-                              group.currency
-                            }`}
-                            className="flex flex-wrap gap-y-4 items-center border-t last-of-type:border-b last-of-type:!mb-4 -mx-6 px-6 py-3"
-                          >
-                            <FormItem className="flex-1 flex flex-row items-start space-x-3 space-y-0">
-                              <FormControl>
-                                <Checkbox
-                                  checked={field.value?.some(
-                                    ({ participant }) => participant === id,
-                                  )}
-                                  onCheckedChange={(checked) => {
-                                    if (readOnly) return
-                                    const options = {
-                                      shouldDirty: true,
-                                      shouldTouch: true,
-                                      shouldValidate: true,
-                                    }
-                                    checked
-                                      ? form.setValue(
-                                          'paidFor',
-                                          [
-                                            ...field.value,
-                                            {
-                                              participant: id,
-                                              shares: '1', // Use string to ensure consistent schema handling
-                                            },
-                                          ] as any,
-                                          options,
-                                        )
-                                      : form.setValue(
-                                          'paidFor',
-                                          field.value?.filter(
-                                            (value) => value.participant !== id,
-                                          ),
-                                          options,
-                                        )
-                                  }}
-                                  disabled={readOnly}
-                                />
-                              </FormControl>
-                              <FormLabel className="text-sm font-normal flex-1">
-                                {name}
-                                {pending && (
-                                  <span className="ml-2 text-xs text-muted-foreground">
-                                    {t('participant.pending')}
-                                  </span>
-                                )}
-                                {field.value?.some(
-                                  ({ participant }) => participant === id,
-                                ) &&
-                                  !form.watch('isReimbursement') && (
-                                    <ParticipantRowAmountPreview
-                                      amount={calculateShare(id, {
-                                        amount: amountAsMinorUnits(
-                                          Number(form.watch('amount')),
-                                          groupCurrency,
-                                        ), // Convert to cents
-                                        paidFor: field.value.map(
-                                          ({ participant, shares }) => ({
-                                            participant: {
-                                              id: participant,
-                                              name: '',
-                                              groupId: '',
-                                            },
-                                            shares:
-                                              form.watch('splitMode') ===
-                                              'BY_PERCENTAGE'
-                                                ? Number(shares) * 100 // Convert percentage to basis points (e.g., 50% -> 5000)
-                                                : form.watch('splitMode') ===
-                                                    'BY_AMOUNT'
-                                                  ? amountAsMinorUnits(
-                                                      shares,
-                                                      groupCurrency,
-                                                    )
-                                                  : shares,
-                                            expenseId: '',
-                                            participantId: '',
-                                          }),
-                                        ),
-                                        splitMode: form.watch('splitMode'),
-                                        isReimbursement:
-                                          form.watch('isReimbursement'),
-                                      })}
-                                      currency={groupCurrency}
-                                    />
-                                  )}
-                              </FormLabel>
-                            </FormItem>
-                            <div className="flex">
-                              {form.getValues().splitMode === 'BY_AMOUNT' &&
-                                !!conversionRequired && (
-                                  <FormField
-                                    name={`paidFor[${field.value.findIndex(
-                                      ({ participant }) => participant === id,
-                                    )}].originalAmount`}
-                                    render={() => {
-                                      const sharesLabel = (
-                                        <span
-                                          className={cn('text-sm', {
-                                            'text-muted': !field.value?.some(
-                                              ({ participant }) =>
-                                                participant === id,
-                                            ),
-                                          })}
-                                        >
-                                          {originalCurrency.symbol}
-                                        </span>
-                                      )
-                                      return (
-                                        <div>
-                                          <div className="flex gap-1 items-center">
-                                            {sharesLabel}
-                                            <FormControl>
-                                              <Input
-                                                key={String(
-                                                  !field.value?.some(
-                                                    ({ participant }) =>
-                                                      participant === id,
-                                                  ),
-                                                )}
-                                                className="text-base w-[80px] -my-2"
-                                                type="text"
-                                                inputMode="decimal"
-                                                disabled={
-                                                  readOnly ||
-                                                  !field.value?.some(
-                                                    ({ participant }) =>
-                                                      participant === id,
-                                                  )
-                                                }
-                                                value={
-                                                  field.value.find(
-                                                    ({ participant }) =>
-                                                      participant === id,
-                                                  )?.originalAmount ?? ''
-                                                }
-                                                onChange={(event) => {
-                                                  const originalAmount = Number(
-                                                    event.target.value,
-                                                  )
-                                                  let convertedAmount = ''
-                                                  if (
-                                                    !Number.isNaN(
-                                                      originalAmount,
-                                                    ) &&
-                                                    exchangeRate.data
-                                                  ) {
-                                                    convertedAmount = (
-                                                      originalAmount *
-                                                      exchangeRate.data
-                                                    ).toFixed(
-                                                      groupCurrency.decimal_digits,
-                                                    )
-                                                  }
-                                                  field.onChange(
-                                                    field.value.map((p) =>
-                                                      p.participant === id
-                                                        ? {
-                                                            participant: id,
-                                                            originalAmount:
-                                                              event.target
-                                                                .value,
-                                                            shares:
-                                                              enforceCurrencyPattern(
-                                                                convertedAmount,
-                                                              ),
-                                                          }
-                                                        : p,
-                                                    ),
-                                                  )
-                                                  setManuallyEditedParticipants(
-                                                    (prev) =>
-                                                      new Set(prev).add(id),
-                                                  )
-                                                }}
-                                                step={
-                                                  10 **
-                                                  -originalCurrency.decimal_digits
-                                                }
-                                              />
-                                            </FormControl>
-                                            <ChevronRight className="h-4 w-4 mx-1 opacity-50" />
-                                          </div>
-                                        </div>
-                                      )
-                                    }}
-                                  />
-                                )}
-                              {form.getValues().splitMode !== 'EVENLY' && (
-                                <FormField
-                                  name={`paidFor[${field.value.findIndex(
-                                    ({ participant }) => participant === id,
-                                  )}].shares`}
-                                  render={() => {
-                                    const sharesLabel = (
-                                      <span
-                                        className={cn('text-sm', {
-                                          'text-muted': !field.value?.some(
-                                            ({ participant }) =>
-                                              participant === id,
-                                          ),
-                                        })}
-                                      >
-                                        {match(form.getValues().splitMode)
-                                          .with('BY_SHARES', () => (
-                                            <>{t('shares')}</>
-                                          ))
-                                          .with('BY_PERCENTAGE', () => <>%</>)
-                                          .with('BY_AMOUNT', () => (
-                                            <>{group.currency}</>
-                                          ))
-                                          .otherwise(() => (
-                                            <></>
-                                          ))}
-                                      </span>
-                                    )
-                                    return (
-                                      <div>
-                                        <div className="flex gap-1 items-center">
-                                          {form.getValues().splitMode ===
-                                            'BY_AMOUNT' && sharesLabel}
-                                          <FormControl>
-                                            <Input
-                                              key={String(
-                                                !field.value?.some(
-                                                  ({ participant }) =>
-                                                    participant === id,
-                                                ),
-                                              )}
-                                              className="text-base w-[80px] -my-2"
-                                              type="text"
-                                              disabled={
-                                                readOnly ||
-                                                !field.value?.some(
-                                                  ({ participant }) =>
-                                                    participant === id,
-                                                )
-                                              }
-                                              value={
-                                                field.value?.find(
-                                                  ({ participant }) =>
-                                                    participant === id,
-                                                )?.shares
-                                              }
-                                              onChange={(event) => {
-                                                field.onChange(
-                                                  field.value.map((p) =>
-                                                    p.participant === id
-                                                      ? {
-                                                          participant: id,
-                                                          shares:
-                                                            enforceCurrencyPattern(
-                                                              event.target
-                                                                .value,
-                                                            ),
-                                                        }
-                                                      : p,
-                                                  ),
-                                                )
-                                                setManuallyEditedParticipants(
-                                                  (prev) =>
-                                                    new Set(prev).add(id),
-                                                )
-                                              }}
-                                              inputMode={
-                                                form.getValues().splitMode ===
-                                                'BY_AMOUNT'
-                                                  ? 'decimal'
-                                                  : 'numeric'
-                                              }
-                                              step={
-                                                form.getValues().splitMode ===
-                                                'BY_AMOUNT'
-                                                  ? 10 **
-                                                    -groupCurrency.decimal_digits
-                                                  : 1
-                                              }
-                                            />
-                                          </FormControl>
-                                          {[
-                                            'BY_SHARES',
-                                            'BY_PERCENTAGE',
-                                          ].includes(
-                                            form.getValues().splitMode,
-                                          ) && sharesLabel}
-                                        </div>
-                                        <FormMessage className="float-right" />
-                                      </div>
-                                    )
-                                  }}
-                                />
-                              )}
-                            </div>
-                          </div>
-                        )
-                      }}
-                    />
-                  ))}
-                  <FormMessage />
-                </FormItem>
-              )}
-            />
-
-            <ParticipantDistributionFooter
-              splitMode={form.watch('splitMode')}
-              targetAmount={
-                form.watch('splitMode') === 'BY_PERCENTAGE'
-                  ? 100
-                  : amountAsMinorUnits(
-                      Number(form.watch('amount')) || 0,
-                      groupCurrency,
-                    )
-              }
-              shares={
-                form.watch('splitMode') === 'BY_AMOUNT'
-                  ? form
-                      .watch('paidFor')
-                      .map((p) =>
-                        amountAsMinorUnits(
-                          Number(p.shares) || 0,
-                          groupCurrency,
-                        ),
-                      )
-                  : form.watch('paidFor').map((p) => Number(p.shares) || 0)
-              }
-              currency={groupCurrency}
-              paidByCount={form.watch('paidFor').length}
-              dataTestId="paid-for-distribution-footer"
-            />
-
-            <Collapsible
-              className="mt-5"
-              defaultOpen={form.getValues().splitMode !== 'EVENLY'}
-            >
-              <CollapsibleTrigger asChild>
-                <Button variant="link" className="-mx-4" disabled={readOnly}>
-                  {t('advancedOptions')}
-                </Button>
-              </CollapsibleTrigger>
-              <CollapsibleContent>
-                <div className="grid sm:grid-cols-2 gap-6 pt-3">
-                  <FormField
-                    control={form.control}
-                    name="splitMode"
-                    render={({ field }) => (
-                      <FormItem>
-                        <FormLabel>{t('SplitModeField.label')}</FormLabel>
-                        <FormControl>
-                          <Select
-                            onValueChange={(value) => {
-                              form.setValue('splitMode', value as any, {
-                                shouldDirty: true,
-                                shouldTouch: true,
-                                shouldValidate: true,
-                              })
-                            }}
-                            defaultValue={field.value}
-                            disabled={readOnly}
-                          >
-                            <SelectTrigger>
-                              <SelectValue />
-                            </SelectTrigger>
-                            <SelectContent>
-                              <SelectItem value="EVENLY">
-                                {t('SplitModeField.evenly')}
-                              </SelectItem>
-                              <SelectItem value="BY_SHARES">
-                                {t('SplitModeField.byShares')}
-                              </SelectItem>
-                              <SelectItem value="BY_PERCENTAGE">
-                                {t('SplitModeField.byPercentage')}
-                              </SelectItem>
-                              <SelectItem value="BY_AMOUNT">
-                                {t('SplitModeField.byAmount')}
-                              </SelectItem>
-                            </SelectContent>
-                          </Select>
-                        </FormControl>
-                        <FormDescription>
-                          {t(`${sExpense}.splitModeDescription`)}
-                        </FormDescription>
-                      </FormItem>
-                    )}
-                  />
-                  <FormField
-                    control={form.control}
-                    name="saveDefaultSplittingOptions"
-                    render={({ field }) => (
-                      <FormItem className="flex flex-row gap-2 items-center space-y-0 pt-2">
-                        <FormControl>
-                          <Checkbox
-                            checked={field.value}
-                            onCheckedChange={field.onChange}
-                            disabled={readOnly}
-                          />
-                        </FormControl>
-                        <div>
-                          <FormLabel>
-                            {t('SplitModeField.saveAsDefault')}
-                          </FormLabel>
-                        </div>
-                      </FormItem>
-                    )}
-                  />
-                </div>
-              </CollapsibleContent>
-            </Collapsible>
-          </CardContent>
-        </Card>
-
-        {runtimeFeatureFlags.enableExpenseDocuments && (
-          <Card className="mt-4">
-            <CardHeader>
-              <CardTitle className="flex justify-between">
-                <span>{t('attachDocuments')}</span>
-              </CardTitle>
-              <CardDescription>
-                {t(`${sExpense}.attachDescription`)}
-              </CardDescription>
-            </CardHeader>
-            <CardContent>
-              <FormField
-                control={form.control}
-                name="documents"
-                render={({ field }) => (
-                  <ExpenseDocumentsInput
-                    documents={field.value}
-                    updateDocuments={field.onChange}
-                    ledgerId={group.ledgerId}
-                    readOnly={readOnly}
-                  />
-                )}
-              />
-            </CardContent>
-          </Card>
-        )}
-
-        {!readOnly && (
-          <div className="flex mt-4 gap-2">
-            <SubmitButton loadingContent={t(isCreate ? 'creating' : 'saving')}>
-              <Save className="w-4 h-4 mr-2" />
-              {t(isCreate ? 'create' : 'save')}
-            </SubmitButton>
-            {!isCreate && onDelete && (
-              <DeletePopup onDelete={() => onDelete()}></DeletePopup>
-            )}
-            <Button variant="ghost" asChild>
-              <Link href={`/groups/${group.id}`}>{t('cancel')}</Link>
-            </Button>
-          </div>
-        )}
-        {readOnly && (
-          <div className="flex mt-4 gap-2">
-            <Button variant="ghost" asChild>
-              <Link href={`/groups/${group.id}`}>{t('cancel')}</Link>
-            </Button>
-          </div>
-        )}
-      </form>
-    </Form>
-  )
-}
-
-function formatDate(date?: Date) {
-  if (!date || isNaN(date as any)) date = new Date()
-  return date.toISOString().substring(0, 10)
-}
+// Re-export from the refactored folder so existing imports keep working.
+export { ExpenseForm } from './expense-form/index'
