+++ apps/web/src/app/groups/[groupId]/expenses/expense-form/participant-pending-label.tsx
@@ -1,0 +1,15 @@
+import { cn } from '@/lib/utils'
+
+export function ParticipantPendingLabel({
+  text,
+  className,
+}: {
+  text: string
+  className?: string
+}) {
+  return (
+    <span className={cn('text-muted-foreground text-xs ml-1', className)}>
+      {text}
+    </span>
+  )
+}
