+++ apps/web/src/app/groups/[groupId]/expenses/expense-form/participant-share-row.tsx
@@ -1,0 +1,60 @@
+import { Checkbox } from '@/components/ui/checkbox'
+import { FormControl, FormItem, FormLabel } from '@/components/ui/form'
+import { cn } from '@/lib/utils'
+import type { ReactNode } from 'react'
+
+export function ParticipantShareRow({
+  participant,
+  checked,
+  onCheckedChange,
+  preview,
+  pendingLabel,
+  shareInput,
+  className,
+  disabled,
+  dataId,
+}: {
+  participant: {
+    id: string
+    name: string
+    pending?: boolean
+    unlinked?: boolean
+  }
+  checked: boolean
+  onCheckedChange: (next: boolean) => void
+  preview?: ReactNode
+  pendingLabel?: ReactNode
+  shareInput?: ReactNode
+  className?: string
+  disabled?: boolean
+  dataId?: string
+}) {
+  return (
+    <div
+      className={cn(
+        'flex flex-wrap gap-y-4 items-center border-t last-of-type:border-b last-of-type:!mb-4 -mx-6 px-6 py-3',
+        className,
+      )}
+      data-id={dataId}
+    >
+      <FormItem className="flex-1 flex flex-row items-start space-x-3 space-y-0">
+        <FormControl>
+          <Checkbox
+            checked={checked}
+            onCheckedChange={(next) => {
+              if (disabled) return
+              onCheckedChange(next as boolean)
+            }}
+            disabled={disabled}
+          />
+        </FormControl>
+        <FormLabel className="text-sm font-normal flex-1">
+          {participant.name}
+          {pendingLabel}
+          {preview}
+        </FormLabel>
+      </FormItem>
+      {shareInput && <div className="flex">{shareInput}</div>}
+    </div>
+  )
+}
